Create a rock, paper scissors game.
The user picks a one player game, two players game, or an automatic game. For the one player game, the player chooses rock, paper, or scissors and the computer will randomly roll and the program calculates who wins. For the two players game, both players pick rock, paper, or scissors and the program determines who wins. For the computer vs. computer game, the user sets a number of games between 10 and 10,000 on the trackbar and the program randomizes the rolls and calculates to total wins. The output is the total win statistics for player 1, player 2, and the ties.